#96 remove shift in multipart optimiser, and also some debug stuff

+++ bouncer/src/repo/core/model/repo_node_utils.h
@@ -284,9 +284,9 @@ static repo_vector_t multiplyMatVec(const std::vector<float> &mat, const repo_ve
 			12 13 14 15
 		*/
 
-		result.x = (double)mat[0] * (double)vec.x + (double)mat[1] * (double)vec.y + (double)mat[2] * (double)vec.z + (double)mat[3];
-		result.y = (double)mat[4] * (double)vec.x + (double)mat[5] * (double)vec.y + (double)mat[6] * (double)vec.z + (double)mat[7];
-		result.z = (double)mat[8] * (double)vec.x + (double)mat[9] * (double)vec.y + (double)mat[10] * (double)vec.z + (double)mat[11];
+		result.x = mat[0] * vec.x + mat[1] * vec.y + mat[2] * vec.z + mat[3];
+		result.y = mat[4] * vec.x + mat[5] * vec.y + mat[6] * vec.z + mat[7];
+		result.z = mat[8] * vec.x + mat[9] * vec.y + mat[10] * vec.z + mat[11];
 
 		float sig = 1e-5;
